chore: promote dashboard to / and drop google-related config
- Move /dashboard page to / (root). Nav link and logo updated.
- Remove google-sync topbar icon and unused .topbar-icon CSS.
- Drop googleapis dependency from package.json + lockfile (-66 packages).
- Remove googleapis/google-auth-library from vite externals.
- Drop NODE_OPTIONS heap workaround in release.yml (no longer needed
  without googleapis in the bundle, reverting the mitigation from #179).
- Remove dead refs to /annotate, /entrenamiento, /google-sync from
  layout.example.svelte (UI demo file).
- Drop ZoneAnnotationRepository usage from /api/invoices/[id] GET response.
- Delete docs/GOOGLE_SETUP.md.
