+++ .github/workflows/release.yml
@@ -31,10 +31,6 @@ jobs:
 
       - name: Build application
         run: npm run build
-        env:
-          # vite build se queda sin heap (default ~4GB) bundleando googleapis + Svelte server.
-          # Hasta que la integración con Google quede out, subimos el heap.
-          NODE_OPTIONS: --max-old-space-size=6144
 
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v3
