+++ client/vite.config.ts
@@ -46,8 +46,6 @@ export default defineConfig(({ mode }) => {
           'pdf-parse',
           'sharp',
           'exceljs',
-          'googleapis',
-          'google-auth-library',
           'tesseract.js',
           'pdf-to-img',
           'pdf-lib',
@@ -64,8 +62,6 @@ export default defineConfig(({ mode }) => {
         'pdf-parse',
         'sharp',
         'exceljs',
-        'googleapis',
-        'google-auth-library',
         'tesseract.js',
         'pdf-to-img',
         'pdf-lib',
